implemented environment object to simplify property passing
