+++ EmberMate/Views/AppView.swift
@@ -9,14 +9,12 @@ import Foundation
 import SwiftUI
 
 struct AppView: View {
-    @ObservedObject var emberMug: EmberMug
-    var appState: AppState
-    @ObservedObject var bluetoothManager: BluetoothManager
+    @EnvironmentObject private var bluetoothManager: BluetoothManager
 
     var body: some View {
         switch bluetoothManager.state {
-        case .connected: MugControlView(emberMug: emberMug, appState: appState)
-        case .disconnected, .connecting: ConnectMugView(bluetoothManager: bluetoothManager)
+        case .connected: MugControlView()
+        case .disconnected, .connecting: ConnectMugView()
         case .reConnecting: Text("Device Lost, Searching...").padding()
         }
     }
